Move gitea endpoints to directory
