+++ src/endpoints/gitea2/mod.rs
@@ -1,0 +1,5 @@
+pub mod get_gitea_endpoints;
+pub mod gitea_remote_repos;
+pub mod gitea_proxy_login;
+pub mod gitea_proxy_logout;
+pub mod get_my_collaborators;
