+++ src/endpoints/mod.rs
@@ -4,10 +4,10 @@ pub mod debug;
 pub mod i18n;
 pub mod sse;
 pub mod navigation;
-pub mod gitea;
 pub mod clients;
 pub(crate) mod error;
 pub mod app_state;
 pub mod content_utils;
 pub mod git2;
 pub mod burrito2;
+pub mod gitea2;
